renamed common package to batch, batch var names to b, and added defer unlock. Modified code comments to make the aggregator functions logic clearer
