+++ batch/batch_test.go
@@ -1,4 +1,4 @@
-package common
+package batch
 
 import (
 	"errors"
