+++ dataset/dataset.go
@@ -11,8 +11,8 @@ import (
 	"sort"
 	"strings"
 
+	"github.com/ONSdigital/dp-api-clients-go/batch"
 	"github.com/ONSdigital/dp-api-clients-go/clientlog"
-	"github.com/ONSdigital/dp-api-clients-go/common"
 	healthcheck "github.com/ONSdigital/dp-api-clients-go/health"
 	health "github.com/ONSdigital/dp-healthcheck/healthcheck"
 	dprequest "github.com/ONSdigital/dp-net/request"
@@ -787,16 +787,19 @@ func (c *Client) GetOptions(ctx context.Context, userAuthToken, serviceAuthToken
 // GetOptionsInBatches retrieves a list of the dimension options in concurrent batches and accumulates the results
 func (c *Client) GetOptionsInBatches(ctx context.Context, userAuthToken, serviceAuthToken, collectionID, id, edition, version, dimension string, batchSize, maxWorkers int) (opts Options, err error) {
 
-	// function to aggregate items. The first received batch will initialise the structure with items with a fixed size
-	// and items are aggregated in order even if batches are received in the wrong order (e.g. due to concurrent calls)
-	var processBatch OptionsBatchProcessor = func(batch Options) (abort bool, err error) {
-		if len(opts.Items) == 0 {
-			opts.TotalCount = batch.TotalCount
-			opts.Items = make([]Option, batch.TotalCount)
-			opts.Count = batch.TotalCount
+	// Function to aggregate items.
+	// For the first received batch, as we have the total count information, will initialise the final structure of items with a fixed size equal to TotalCount.
+	// This serves two purposes:
+	//   - We can guarantee, even with concurrent calls, that values are returned in the same order that the API defines, by offsetting the index.
+	//   - We do a single memory allocation for the final array, making the code more memory efficient.
+	var processBatch OptionsBatchProcessor = func(b Options) (abort bool, err error) {
+		if len(opts.Items) == 0 { // first batch response being handled
+			opts.TotalCount = b.TotalCount
+			opts.Items = make([]Option, b.TotalCount)
+			opts.Count = b.TotalCount
 		}
-		for i := 0; i < len(batch.Items); i++ {
-			opts.Items[i+batch.Offset] = batch.Items[i]
+		for i := 0; i < len(b.Items); i++ {
+			opts.Items[i+b.Offset] = b.Items[i]
 		}
 		return false, nil
 	}
@@ -815,31 +818,30 @@ func (c *Client) GetOptionsBatchProcess(ctx context.Context, userAuthToken, serv
 	// for each batch, obtain the dimensions starting at the provided offset, with a batch size limit,
 	// or the subste of IDs according to the provided offset, if a list of optionIDs was provided
 	batchGetter := func(offset int) (interface{}, int, error) {
-		var batch Options
 
 		// if a list of IDs is provided, then obtain only the options for that list in batches.
 		if optionIDs != nil {
-			batchEnd := common.Min(len(*optionIDs), offset+batchSize)
+			batchEnd := batch.Min(len(*optionIDs), offset+batchSize)
 			batchOptionIDs := (*optionIDs)[offset:batchEnd]
-			batch, err := c.GetOptions(ctx, userAuthToken, serviceAuthToken, collectionID, id, edition, version, dimension, QueryParams{IDs: batchOptionIDs})
-			return batch, len(*optionIDs), err
+			b, err := c.GetOptions(ctx, userAuthToken, serviceAuthToken, collectionID, id, edition, version, dimension, QueryParams{IDs: batchOptionIDs})
+			return b, len(*optionIDs), err
 		}
 
 		// otherwise obtain all the options in batches.
-		batch, err := c.GetOptions(ctx, userAuthToken, serviceAuthToken, collectionID, id, edition, version, dimension, QueryParams{Offset: offset, Limit: batchSize})
-		return batch, batch.TotalCount, err
+		b, err := c.GetOptions(ctx, userAuthToken, serviceAuthToken, collectionID, id, edition, version, dimension, QueryParams{Offset: offset, Limit: batchSize})
+		return b, b.TotalCount, err
 	}
 
 	// cast and process the batch according to the provided method
-	batchProcessor := func(batch interface{}) (abort bool, err error) {
-		v, ok := batch.(Options)
+	batchProcessor := func(b interface{}) (abort bool, err error) {
+		v, ok := b.(Options)
 		if !ok {
 			return true, errors.New("wrong type")
 		}
 		return processBatch(v)
 	}
 
-	return common.ProcessInConcurrentBatches(batchGetter, batchProcessor, batchSize, maxWorkers)
+	return batch.ProcessInConcurrentBatches(batchGetter, batchProcessor, batchSize, maxWorkers)
 }
 
 // NewDatasetAPIResponse creates an error response, optionally adding body to e when status is 404
