+++ filter/filter.go
@@ -10,8 +10,8 @@ import (
 	"net/http"
 	"strconv"
 
+	"github.com/ONSdigital/dp-api-clients-go/batch"
 	"github.com/ONSdigital/dp-api-clients-go/clientlog"
-	"github.com/ONSdigital/dp-api-clients-go/common"
 	"github.com/ONSdigital/dp-api-clients-go/headers"
 	healthcheck "github.com/ONSdigital/dp-api-clients-go/health"
 	health "github.com/ONSdigital/dp-healthcheck/healthcheck"
@@ -286,16 +286,19 @@ func (c *Client) GetDimensionOptionsBytes(ctx context.Context, userAuthToken, se
 // GetDimensionOptionsInBatches retrieves a list of the dimension options in concurrent batches and accumulates the results
 func (c *Client) GetDimensionOptionsInBatches(ctx context.Context, userAuthToken, serviceAuthToken, collectionID, filterID, name string, batchSize, maxWorkers int) (opts DimensionOptions, err error) {
 
-	// function to aggregate items. The first received batch will initialise the structure with items with a fixed size
-	// and items are aggregated in order even if batches are received in the wrong order (e.g. due to concurrent calls)
-	var processBatch DimensionOptionsBatchProcessor = func(batch DimensionOptions) (abort bool, err error) {
+	// Function to aggregate items.
+	// For the first received batch, as we have the total count information, will initialise the final structure of items with a fixed size equal to TotalCount.
+	// This serves two purposes:
+	//   - We can guarantee, even with concurrent calls, that values are returned in the same order that the API defines, by offsetting the index.
+	//   - We do a single memory allocation for the final array, making the code more memory efficient.
+	var processBatch DimensionOptionsBatchProcessor = func(b DimensionOptions) (abort bool, err error) {
 		if len(opts.Items) == 0 {
-			opts.TotalCount = batch.TotalCount
-			opts.Items = make([]DimensionOption, batch.TotalCount)
-			opts.Count = batch.TotalCount
+			opts.TotalCount = b.TotalCount
+			opts.Items = make([]DimensionOption, b.TotalCount)
+			opts.Count = b.TotalCount
 		}
-		for i := 0; i < len(batch.Items); i++ {
-			opts.Items[i+batch.Offset] = batch.Items[i]
+		for i := 0; i < len(b.Items); i++ {
+			opts.Items[i+b.Offset] = b.Items[i]
 		}
 		return false, nil
 	}
@@ -312,20 +315,20 @@ func (c *Client) GetDimensionOptionsBatchProcess(ctx context.Context, userAuthTo
 
 	// for each batch, obtain the dimensions starting at the provided offset, with a batch size limit
 	batchGetter := func(offset int) (interface{}, int, error) {
-		batch, err := c.GetDimensionOptions(ctx, userAuthToken, serviceAuthToken, collectionID, filterID, name, QueryParams{Offset: offset, Limit: batchSize})
-		return batch, batch.TotalCount, err
+		b, err := c.GetDimensionOptions(ctx, userAuthToken, serviceAuthToken, collectionID, filterID, name, QueryParams{Offset: offset, Limit: batchSize})
+		return b, b.TotalCount, err
 	}
 
 	// cast and process the batch according to the provided method
-	batchProcessor := func(batch interface{}) (abort bool, err error) {
-		v, ok := batch.(DimensionOptions)
+	batchProcessor := func(b interface{}) (abort bool, err error) {
+		v, ok := b.(DimensionOptions)
 		if !ok {
 			return true, errors.New("wrong type")
 		}
 		return processBatch(v)
 	}
 
-	return common.ProcessInConcurrentBatches(batchGetter, batchProcessor, batchSize, maxWorkers)
+	return batch.ProcessInConcurrentBatches(batchGetter, batchProcessor, batchSize, maxWorkers)
 }
 
 // CreateBlueprint creates a filter blueprint and returns the associated filterID
